requirements.txt
discord.py[voice]>=2.3.2
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
PyNaCl>=1.5.0